validacija unosa back

+++ backend-app-myhouse/src/main/java/antihackerman/model/IntervalDevice.java
@@ -22,13 +22,13 @@ public class IntervalDevice extends Device{
     private String valueDefinition;
 
     @Column(name = "max_value", nullable = false)
-    private Double maxValue;
+    private Long maxValue;
 
     @Column(name = "min_value", nullable = false)
-    private Double minValue;
+    private Long minValue;
 
     @Builder
-    public IntervalDevice(Integer id, boolean deleted, String name, String filePath, String filter, Integer readIntervalMils, DeviceType type, RealEstate realestate, Set<DeviceAlarm> alarms, String valueDefinition, Double maxValue, Double minValue) {
+    public IntervalDevice(Integer id, boolean deleted, String name, String filePath, String filter, Integer readIntervalMils, DeviceType type, RealEstate realestate, Set<DeviceAlarm> alarms, String valueDefinition, Long maxValue, Long minValue) {
         super(id, deleted, name, filePath, filter, readIntervalMils, type, realestate, alarms);
         this.valueDefinition = valueDefinition;
         this.maxValue = maxValue;
